@@ -1,6 +1,6 @@
 {
   "name": "ChessColate",
-  "version": "1.2.2",
+  "version": "1.2.3",
   "author": "Jheison Alexander Alzate",
   "homepage": "https://chesscolate.com/",
   "scripts": {
@@ -19,14 +19,13 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^16.2.12",
-    "@angular/core": "^16.2.12",
-    "@angular/forms": "^16.2.12",
-    "@angular/platform-browser": "^16.2.12",
-    "@angular/platform-browser-dynamic": "^16.2.12",
-    "@angular/platform-server": "^16.2.12",
-    "@angular/router": "^16.2.12",
-    "@angular/service-worker": "^16.2.12",
+    "@angular/common": "^17.3.12",
+    "@angular/core": "^17.3.12",
+    "@angular/forms": "^17.3.12",
+    "@angular/platform-browser": "^17.3.12",
+    "@angular/platform-browser-dynamic": "^17.3.12",
+    "@angular/router": "^17.3.12",
+    "@angular/service-worker": "^17.3.12",
     "@capacitor-firebase/authentication": "^6.0.0",
     "@capacitor/android": "6.1.0",
     "@capacitor/app": "6.0.0",
@@ -40,21 +39,17 @@
     "@capacitor/push-notifications": "^6.0.0",
     "@capacitor/status-bar": "6.0.0",
     "@ionic/angular": "^7.0.3",
-    "@ionic/angular-server": "^7.8.1",
     "@ionic/core": "^7.0.3",
     "@ionic/storage-angular": "^4.0.0",
-    "@ngneat/transloco": "^4.0.0",
-    "@ngrx/component-store": "^16.3.0",
-    "@ngrx/effects": "^16.3.0",
-    "@ngrx/entity": "^16.3.0",
-    "@ngrx/router-store": "^16.3.0",
-    "@ngrx/store": "^16.3.0",
-    "@ngrx/store-devtools": "^16.3.0",
-    "@nguniversal/express-engine": "^16.2.0",
-    "@sentry/angular-ivy": "^7.54.0",
-    "@sentry/capacitor": "^0.18.0",
+    "@ngrx/component-store": "^17.2.0",
+    "@ngrx/effects": "^17.2.0",
+    "@ngrx/entity": "^17.2.0",
+    "@ngrx/router-store": "^17.2.0",
+    "@ngrx/store": "^17.2.0",
+    "@ngrx/store-devtools": "^17.2.0",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
     "@types/lodash": "^4.14.186",
-    "angular-google-tag-manager": "^1.8.0",
     "animate.css": "^4.1.1",
     "chart.js": "^3.8.0",
     "chartjs-adapter-date-fns": "^3.0.0",
@@ -62,37 +57,31 @@
     "chess.js": "^0.12.1",
     "cm-chessboard": "^7.7.1",
     "date-fns": "^2.28.0",
-    "express": "^4.15.2",
     "firebase": "^10.12.0",
     "howler": "^2.2.3",
     "lodash": "^4.17.21",
     "rxjs": "~7.5.5",
     "tslib": "^2.2.0",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^16.0.1",
-    "@angular-devkit/build-angular": "^16.2.12",
+    "@angular-devkit/build-angular": "^17.3.8",
     "@angular-eslint/builder": "~15.1.0",
     "@angular-eslint/eslint-plugin": "~15.1.0",
     "@angular-eslint/eslint-plugin-template": "~15.1.0",
     "@angular-eslint/template-parser": "~15.1.0",
-    "@angular/cli": "^16.2.12",
-    "@angular/compiler": "^16.2.12",
-    "@angular/compiler-cli": "^16.2.12",
-    "@angular/language-service": "^16.2.12",
+    "@angular/cli": "^17.3.8",
+    "@angular/compiler": "^17.3.12",
+    "@angular/compiler-cli": "^17.3.12",
+    "@angular/language-service": "^17.3.12",
     "@capacitor/cli": "^6.1.0",
     "@ionic/angular-toolkit": "^5.0.0",
-    "@ngrx/schematics": "^16.3.0",
-    "@nguniversal/builders": "^16.2.0",
-    "@types/express": "^4.17.0",
+    "@ngrx/schematics": "^17.2.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "@typescript-eslint/eslint-plugin": "5.3.0",
     "@typescript-eslint/parser": "5.3.0",
-    "brotli-webpack-plugin": "^1.1.0",
-    "compression-webpack-plugin": "^11.1.0",
     "eslint": "^7.6.0",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-jsdoc": "30.7.6",
@@ -108,7 +97,7 @@
     "protractor": "~7.0.0",
     "rimraf": "^5.0.7",
     "ts-node": "~8.3.0",
-    "typescript": "~5.1.6"
+    "typescript": "~5.4.5"
   },
   "description": "Chess tactic platform"
 }
