@@ -2,7 +2,11 @@ import { APP_INITIALIZER, ErrorHandler, NgModule, isDevMode } from '@angular/cor
 import { BrowserModule } from '@angular/platform-browser';
 import { Router, RouteReuseStrategy } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HttpClient } from '@angular/common/http';
+
+
+import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
 import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
 import { IonicStorageModule } from '@ionic/storage-angular';
@@ -27,31 +31,18 @@ import { CustomRouterStateSerializer } from '@redux/states/router.state';
 import * as fromEffects from '@redux/effects';
 import * as fromGuards from '@guards/index';
 
-import * as Sentry from '@sentry/capacitor';
-import * as SentryAngular from '@sentry/angular-ivy';
-
-Sentry.init(
-  {
-    dsn: 'https://4716d5ae6b19528674465cfafd357307@o4507017740681216.ingest.us.sentry.io/4507019628314624',
-    // To set your release and dist versions
-    release: 'chesscolate@' + environment.version,
-    dist: '1',
-    // Set tracesSampleRate to 1.0 to capture 100% of transactions for performance monitoring.
-    // We recommend adjusting this value in production.
-    tracesSampleRate: environment.production ? 0.25 : 1,
-    environment: environment.production ? 'production' : 'development',
-    ignoreErrors: [
-    ],
-  },
-  // Forward the init method from @sentry/angular
-  SentryAngular.init
-);
 
 
-import { TranslocoRootModule } from './transloco-root.module';
+
+
+// AoT requires an exported function for factories
+// eslint-disable-next-line prefer-arrow/prefer-arrow-functions, @typescript-eslint/naming-convention
+export function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http);
+}
+
 import { ServiceWorkerModule } from '@angular/service-worker';
 
-import { GoogleTagManagerModule } from 'angular-google-tag-manager';
 
 
 
@@ -77,6 +68,14 @@ if (environment.production) {
     IonicModule.forRoot(),
     IonicStorageModule.forRoot(),
     AppRoutingModule,
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        deps: [HttpClient]
+      },
+      defaultLanguage: 'en'
+    }),
     /* NGRX */
     StoreRouterConnectingModule.forRoot({
       serializer: CustomRouterStateSerializer
@@ -86,7 +85,6 @@ if (environment.production) {
     EffectsModule.forRoot(fromEffects.EFFECTS),
     // SocketIoModule.forRoot(config),
     HttpClientModule,
-    TranslocoRootModule,
     FormsModule,
     ReactiveFormsModule,
     SharedModule,
@@ -95,28 +93,11 @@ if (environment.production) {
       // Register the ServiceWorker as soon as the application is stable
       // or after 30 seconds (whichever comes first).
       registrationStrategy: 'registerWhenStable:30000'
-    }),
-    GoogleTagManagerModule.forRoot({ id: 'G-20ME9L8R51' })
-
+    })
   ],
   providers: [
     PROVIDERS,
-    { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
-    {
-      provide: ErrorHandler,
-      // Attach the Sentry ErrorHandler
-      useValue: SentryAngular.createErrorHandler(),
-    },
-    {
-      provide: SentryAngular.TraceService,
-      deps: [Router],
-    },
-    {
-      provide: APP_INITIALIZER,
-      useFactory: () => () => { },
-      deps: [SentryAngular.TraceService],
-      multi: true,
-    },
+    { provide: RouteReuseStrategy, useClass: IonicRouteStrategy }
   ],
   bootstrap: [AppComponent]
 })
