@@ -1,9 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { NavController, LoadingController, ModalController } from '@ionic/angular';
-import { TranslocoService } from '@ngneat/transloco';
+import { TranslateService } from '@ngx-translate/core';
 
 import { Meta } from '@angular/platform-browser';
-import { GoogleTagManagerService } from 'angular-google-tag-manager';
 
 import { Plan, Block, PlanTypes } from '@models/plan.model';
 import { PlanService } from '@services/plan.service';
@@ -35,9 +34,8 @@ export class TrainingMenuComponent implements OnInit {
     private loadingController: LoadingController,
     private profileService: ProfileService,
     private meta: Meta,
-    private googleTagManagerService: GoogleTagManagerService,
     private modalController: ModalController,
-    private translocoService: TranslocoService
+    private translateService: TranslateService
   ) { }
 
   ngOnInit() {
@@ -76,7 +74,6 @@ export class TrainingMenuComponent implements OnInit {
 
     const newPlan: Plan = await this.planService.newPlan(blocks, planType, option * 60);
     this.closeLoading();
-    this.googleTagManagerService.pushTag({ event: 'newPlan', planType, planTime: option });
     this.goTo('/puzzles/training');
 
   }
@@ -84,7 +81,7 @@ export class TrainingMenuComponent implements OnInit {
 
   async showLoading() {
     this.loader = await this.loadingController.create({
-      message: this.translocoService.translate('loadingPuzzles'),
+      message: this.translateService.instant('loadingPuzzles'),
     });
 
     this.loader.present();
@@ -107,7 +104,6 @@ export class TrainingMenuComponent implements OnInit {
     });
 
     await modal.present();
-    // this.googleTagManagerService.pushTag({ event: 'showChart', planType: planType });
   }
 
 
