@@ -2,13 +2,12 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { NavController, ModalController } from '@ionic/angular';
 import { Meta } from '@angular/platform-browser';
 
-import { GoogleTagManagerService } from 'angular-google-tag-manager';
 
 
 import { Subject, interval } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 
-import { TranslocoService } from '@ngneat/transloco';
+import { TranslateService } from '@ngx-translate/core';
 
 /**
  * Funcionalidad: que al mostrar la solución lo haga que cada jugada deje flechas,  después
@@ -67,7 +66,7 @@ export class TrainingComponent implements OnInit {
 
   profile: Profile;
 
-  currentLanguage = this.translocoService.getActiveLang();
+  currentLanguage = this.translateService.currentLang;
 
   constructor(
     private planService: PlanService,
@@ -77,10 +76,8 @@ export class TrainingComponent implements OnInit {
     private modalController: ModalController,
     public appService: AppService,
     private soundsService: SoundsService,
-    private translocoService: TranslocoService,
-    private meta: Meta,
-    private googleTagManagerService: GoogleTagManagerService
-  ) {
+    private translateService: TranslateService,
+    private meta: Meta) {
   }
 
   ngOnInit() {
@@ -387,8 +384,7 @@ export class TrainingComponent implements OnInit {
       // console.log('Plan finalizado ', JSON.stringify(this.plan));
       this.planService.requestSavePlanAction(this.plan);
     }
-    this.googleTagManagerService.pushTag({ event: 'endPlan', planType: this.plan.planType });
-
+    // TODO: Track end plan
   }
 
 
@@ -416,6 +412,6 @@ export class TrainingComponent implements OnInit {
     });
 
     await modal.present();
-    // this.googleTagManagerService.pushTag({ event: 'showChart', planType: planType });
+
   }
 }
